mysql_num_fields returns an unsigned int

+++ ext/mysql2/result.h
@@ -8,7 +8,7 @@ typedef struct {
   VALUE fields;
   VALUE rows;
   VALUE encoding;
-  long numberOfFields;
+  unsigned int numberOfFields;
   unsigned long numberOfRows;
   unsigned long lastRowProcessed;
   char resultFreed;
